*minor classes refactoring *added physical directories for classes in "phantasm-core" *added own "main" function in "phm-core-tester"
