+++ phantasm/src/phantasm-core/image_saver/image_saver.cpp
@@ -1,0 +1,12 @@
+#include <fstream>
+
+#include "image_saver.h"
+
+void SaveImageToJpeg (const IMAGE_STORAGE & img, std::string fileName)
+{
+   std::ofstream out(fileName);
+
+
+
+   out.close();
+}
