+++ phantasm/src/phantasm-core/image_storage/image_storage.cpp
@@ -1,5 +1,3 @@
-#include <fstream>
-
 #include "image_storage.h"
 
 IMAGE_STORAGE::IMAGE_STORAGE (void) : w(0), h(0), data(NULL)
@@ -10,13 +8,4 @@ IMAGE_STORAGE::IMAGE_STORAGE (void) : w(0), h(0), data(NULL)
 IMAGE_STORAGE::~IMAGE_STORAGE (void)
 {
 
-}
-
-void IMAGE_STORAGE::SaveToJpeg (std::string fileName)
-{
-   std::ofstream out(fileName);
-
-
-
-   out.close();
 }
