+++ phantasm/src/phantasm-core/image_storage/image_storage.h
@@ -2,14 +2,13 @@
 #ifndef _IMAGE_H_
 #define _IMAGE_H_
 
-#include "phm_core.h"
-#include "common.h"
+#include "..\phm_core.h"
+#include "..\common.h"
 
 class PHM_CORE_API IMAGE_STORAGE {
 public:
    IMAGE_STORAGE (void);
    ~IMAGE_STORAGE (void);
-   void SaveToJpeg (std::string fileName);
 private:
    int w;
    int h;
