+++ phantasm/src/phantasm-core/math/vec.cpp
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "..\common.h"
 
 #include "vecmat.h"
