+++ phantasm/src/phantasm-core/objects/object.cpp
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "..\common.h"
 
 #include "object.h"
