+++ phantasm/src/phantasm-core/phm_core_export.h
@@ -3,12 +3,13 @@
 #define _PMH_CORE_EXPORT_H_
 
 #include "common.h"
-#include "object.h"
-#include "sphere.h"
-#include "scene.h"
-#include "vecmat.h"
-#include "image_storage.h"
-#include "scene_loader.h"
-#include "tracer.h"
+#include "objects\object.h"
+#include "objects\sphere.h"
+#include "scene\scene.h"
+#include "math\vecmat.h"
+#include "image_storage\image_storage.h"
+#include "image_saver\image_saver.h"
+#include "scene_loader\scene_loader.h"
+#include "tracer\tracer.h"
 
 #endif // _PMH_CORE_EXPORT_H_
