+++ phantasm/src/phantasm-core/scene/scene.cpp
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "..\common.h"
 
 #include <vector>
