+++ phantasm/src/phantasm-core/tracer/tracer.cpp
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "..\common.h"
 
 #include "tracer.h"
